requirements.txt
streamlit==1.28.0
langchain==0.1.0
langchain-groq==0.1.0
groq==0.4.0
easyocr==1.7.0
Pillow==10.1.0
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
pydantic==2.5.0